requirements.txt
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
chart-studio==1.1.0
click==8.0.1
coverage==5.5
cssselect==1.1.0
cycler==0.10.0
descartes==1.1.0
Django==3.2.7
django-cors==0.1
django-cors-headers==3.8.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
geographiclib==1.52
geopy==2.2.0
geotext==0.4.0
google==3.0.0
gunicorn==20.1.0
html2text==2020.1.16
idna==3.2
iniconfig==1.1.1
jieba3k==0.35.1
joblib==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
pluggy==1.0.0
pprintpp==0.4.0
py==1.10.0
pycountry==20.7.3
pycountry-convert==0.7.2
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-mock==3.6.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
repoze.lru==0.7
requests==2.26.0
requests-file==1.5.1
retrying==1.3.3
sgmllib3k==1.0.0
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tenacity==8.0.1
tinysegmenter==0.3
tldextract==3.1.2
toml==0.10.2
tqdm==4.62.2
urllib3==1.26.6
wordcloud==1.8.1
